feat(layout): static public root layout with ConsentGate bootstrap
Remove cookies(), headers(), and draftMode() from the public root layout
so marketing routes can static-render. Delegate html/head/body to
SiteStaticShell, isolate CSP nonce reads for JSON-LD in a small dynamic
child, wire ConsentGate for consent/GTM/Analytics, and drop force-dynamic
from blog section components.
